% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{find_best_k_cross_entropy}
\alias{find_best_k_cross_entropy}
\title{Find Optimal K (Cross-Entropy)
(Task 5)}
\usage{
find_best_k_cross_entropy(train_data, valid_data, k_values = 1:30)
}
\arguments{
\item{train_data}{Training data.}

\item{valid_data}{Validation data.}

\item{k_values}{Vector of K values.}
}
\value{
List with results data frame and best K.
}
\description{
Find Optimal K (Cross-Entropy)
(Task 5)
}
