% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{find_hardest_easiest_8s}
\alias{find_hardest_easiest_8s}
\title{Find Hardest and Easiest Digit Examples
(Task 3)}
\usage{
find_hardest_easiest_8s(train_data, k = 30)
}
\arguments{
\item{train_data}{Training data.}

\item{k}{Number of neighbors (default 30).}
}
\value{
List with hardest (3) and easiest (2) examples.
}
\description{
Find Hardest and Easiest Digit Examples
(Task 3)
}
