% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{optdigits}
\alias{optdigits}
\title{Handwritten Digits Data (Optdigits)}
\format{
A data frame with 5620 rows and 65 variables.
The first 64 columns (named V1, V2, ..., V64) are pixel counts (integers 0-16).
The 65th column (named Digit) is the target factor (digits 0-9).
}
\source{
UCI Machine Learning Repository (Optical Recognition of Handwritten Digits)
}
\usage{
optdigits
}
\description{
A dataset containing normalized bitmaps of handwritten digits.
The data was derived from 32x32 bitmaps, which were then
divided into 4x4 blocks, and the number of "on" pixels
was counted in each block. This resulted in an 8x8 matrix.
}
\keyword{datasets}
