% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot Confusion Matrix Heatmap
(Task 2 Helper)}
\usage{
plot_confusion_matrix(cm, title = "Confusion Matrix")
}
\arguments{
\item{cm}{Confusion matrix.}

\item{title}{Plot subtitle.}
}
\value{
ggplot object.
}
\description{
Plot Confusion Matrix Heatmap
(Task 2 Helper)
}
