% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{plot_digit}
\alias{plot_digit}
\title{Plot a Single Digit
(Helper function for Task 3)}
\usage{
plot_digit(digit_row, title = "")
}
\arguments{
\item{digit_row}{A single row from the optdigits data frame.}

\item{title}{A title for the plot.}
}
\value{
A ggplot object.
}
\description{
Plot a Single Digit
(Helper function for Task 3)
}
