% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{plot_knn_errors}
\alias{plot_knn_errors}
\title{Plot KNN Misclassification Error Curve
(Task 4 Helper)}
\usage{
plot_knn_errors(k_results_df, best_k)
}
\arguments{
\item{k_results_df}{Results data frame.}

\item{best_k}{Optimal K value.}
}
\value{
ggplot object.
}
\description{
Plot KNN Misclassification Error Curve
(Task 4 Helper)
}
