% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{split_optdigits_data}
\alias{split_optdigits_data}
\title{Split the Optdigits Dataset
(Fulfills Task 1)}
\usage{
split_optdigits_data(data, seed = 12345)
}
\arguments{
\item{data}{The complete data frame.}

\item{seed}{A random seed.}
}
\value{
A list containing train_data, valid_data, and test_data.
}
\description{
Split the Optdigits Dataset
(Fulfills Task 1)
}
