% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{evaluate_k_30}
\alias{evaluate_k_30}
\title{Evaluate a K=30 Classifier
(Fulfills Task 2)}
\usage{
evaluate_k_30(train_data, test_data)
}
\arguments{
\item{train_data}{The training dataset.}

\item{test_data}{The test dataset.}
}
\value{
A list containing confusion matrices and error rates.
}
\description{
Evaluate a K=30 Classifier
(Fulfills Task 2)
}
