% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{find_best_k_misclassification}
\alias{find_best_k_misclassification}
\title{Find Optimal K (Misclassification)
(Fulfills Task 4)}
\usage{
find_best_k_misclassification(train_data, valid_data, k_values = 1:30)
}
\arguments{
\item{train_data}{The training dataset.}

\item{valid_data}{The validation dataset.}

\item{k_values}{An integer vector.}
}
\value{
A list containing \code{results_df} and \code{best_k}.
}
\description{
Find Optimal K (Misclassification)
(Fulfills Task 4)
}
