% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{find_hardest_easiest_8s}
\alias{find_hardest_easiest_8s}
\title{Find Hardest and Easiest Digit Examples
(Fulfills Task 3)}
\usage{
find_hardest_easiest_8s(train_data, best_k)
}
\arguments{
\item{train_data}{The training dataset.}

\item{best_k}{The optimal K value (from Task 4).}
}
\value{
A list containing \code{hardest} (3 examples) and \code{easiest} (2 examples).
}
\description{
Find Hardest and Easiest Digit Examples
(Fulfills Task 3)
}
