% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{plot_confusion_matrix}
\alias{plot_confusion_matrix}
\title{Plot a Confusion Matrix Heatmap
(Helper function for Task 2)}
\usage{
plot_confusion_matrix(cm, title = "Confusion Matrix")
}
\arguments{
\item{cm}{A confusion matrix object (from \code{table()}).}

\item{title}{A string for the plot subtitle.}
}
\value{
A ggplot object.
}
\description{
Plot a Confusion Matrix Heatmap
(Helper function for Task 2)
}
