% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{plot_digit}
\alias{plot_digit}
\title{Plot a Single Digit
(Task 3 Helper)}
\usage{
plot_digit(digit_row, title = "")
}
\arguments{
\item{digit_row}{Single row from data.}

\item{title}{Plot title.}
}
\value{
ggplot object.
}
\description{
Plot a Single Digit
(Task 3 Helper)
}
