% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{plot_knn_errors}
\alias{plot_knn_errors}
\title{Plot KNN Misclassification Error Curve
(Helper function for Task 4)}
\usage{
plot_knn_errors(k_results_df, best_k)
}
\arguments{
\item{k_results_df}{The data frame from \code{find_best_k_misclassification}.}

\item{best_k}{The optimal K value.}
}
\value{
A ggplot object.
}
\description{
Plot KNN Misclassification Error Curve
(Helper function for Task 4)
}
