% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_assignment_functions.R
\name{split_optdigits_data}
\alias{split_optdigits_data}
\title{Split the Optdigits Dataset
(Task 1)}
\usage{
split_optdigits_data(data, seed = 12345)
}
\arguments{
\item{data}{Data frame.}

\item{seed}{Random seed.}
}
\value{
List with train, valid, and test sets.
}
\description{
Split the Optdigits Dataset
(Task 1)
}
